requirements.prod.txt
# Production dependencies for MLB Statistics Analysis System
# For web application deployment only (minimal footprint)

# Core dependencies
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
tenacity>=8.2.0,<9.0.0

# Web framework
streamlit>=1.28.0,<2.0.0
plotly>=5.17.0,<6.0.0

# Optional: Error monitoring (recommended for production)
# Uncomment to enable Sentry integration
# sentry-sdk>=1.40.0,<2.0.0

# Optional: Performance monitoring
# Uncomment for advanced metrics
# prometheus-client>=0.19.0,<1.0.0
